media: i2c: max_des: make link_frequency 64 bit

+++ drivers/media/i2c/maxim-serdes/max_des.h
@@ -80,7 +80,7 @@ struct max_des_pipe {
 
 struct max_des_phy {
 	unsigned int index;
-	unsigned int link_frequency;
+	u64 link_frequency;
 	struct v4l2_mbus_config_mipi_csi2 mipi;
 	bool alt_mem_map8;
 	bool alt2_mem_map8;
